FEAT Macros are working!!!
It's not your standard PCRE yet, but it is possible to express the
PCRE expression

     /AB(CC|DDDD)E*F/

in code without having to struggle with the borrow checker, something
you can see happening a lot in the first few test functions.  Instead,
you can import the `re!` macro and type it as:

`let pattern = re!{cat { tok { 'A' },
                         tok { 'B' },
                         alt { cat { tok { 'C' }, tok { 'C' } },
                               cat { tok { 'D' }, tok { 'D' }, tok { 'D' }, tok { 'D' } } },
                         rep { tok { 'E' } },
                         tok { 'F' } } };`

... and the macro will properly scope all the borrows and assignments,
alleviating one of the hardest issues with this library.

One thing I'd like to do is be able to replace all instances of `tok {
'A' }` with just `'A'`.

The macro takes a type argument as an alternative first, so it
is possible to define it as something other than `char`.  See the
`mildly_complex_macro` and `mildly_complex_untyped_macro` for examples.