FEAT Final version of the macro system (at this time)
I had to remove the convenience features of having the expression
be identifiable without the `tok{ }` wrapper, but no go; it became
too complicated when I started working with multiples.  I had a
choice between living with the `tok{ }` thing and not having a
more-than-two cons operator, meaning `cat{ a, b, c }` would have
to be written `cat{a, cat { b, c }}`, and not that many people would
put up with that much lisp.  I decided the more-than-two cons
operator was friendlier.

Anyway, this concludes the first pass of the macros system. Go me.