REFACTOR Moved everything into a "Node."
Refactored the Brzop() to just be an enum, and made the Node the heart
of the system.  It keeps a car and cdr, plus a pointer to its real
derivative and a copy of the token, in case the derivative needs
calculating.  I feel like an "Ah-hah!" moment will happen at any moment
and the meaning of the sixth conditional will make sense to me:

6. When the derivative of the node's children are calculated,
we update the output node (?) such that it's children point to
these derivatives.  (Adams et. al. 2015, page 4.)