FEAT Laziness.  I think.
Refactor the code to put laziness into the operation, such that
when a derivative is demanded, if the previous operation left it
"lazy" then we calculate it by-need.  I'm not entirely convinced
this is correct, but it doesn't seem to be broken at this point;
all the tests are passing, but I won't actually know until I can
replace the Kleene star handler with an expression like this:
     L = (L ◦ c) ∪ c

Renamed "builder" to "language," although that's a bit of a
misnomer still, but it's the best I've been able to come up
with.

I'm still not entirely comfortable with the
traits I've got going here.  There's some discussion (see
https://www.reddit.com/r/rust/comments/9v2tof/lots_of_long_where_clauses_is_there_a_better_way/
for discussion) but at least, for now, this works.