REFACTOR Separate Grammar from front-end API.
The grammar requires trackers for node changes and memoization, but
the front-end API does not.  Refactor the code to separate these
two features and we get automatic re-initialization of the trackers
for the cost of instantiation.  This is much less likely to be a
source of bugs than attempting to clear the trackers before a run.

This also makes the API independent of features added to the grammar.

All tests passing.