FEAT Parsesets broken out into their own file.
ParseSets needed their own implementation to cover the various kinds
of internal permutations Might & Darais listed in their examples and
papers.  I'm not really happy with this, as it puts control over a
major feature in a different file, but it's the best I can do to isolate
the optimizations in a way that makes them coherent to my brain.