FEAT Adams's optimizations, part one.
This code implements all of the optimizations found in _On The
Complexity and Performance of Parsing With Derivatives_ (Adams, et. al., 2015),
although at the moment it does so poorly.  Several of the tests are failing,
and several others have not been realized properly.  But that it compiles
*at all* is a testament to Rust and a joy to behold.

TODO: Write tests that examine each optimization in turn, especially those
that require reductions.

TODO: Consider splitting the reductions into an internal and external
pass, one that will allow the user to specify the data types that the
parse-null pass will extract, and then redefine the internal ones (the
tree-balancing ones) as data types with a common operator name, rather
than as closures.  Leave closures to the hackers.

Optionally: Consider passing in reductions as an object with named
fields, and use an indirect call-by-name syntax to build the parse
tree *inside* the object, and return that instead.  Ugly.  Read up on
inversion of control.
,
although at the moment it does so poorly.  Several of the tests are failing,
and several others have not been realized properly.  But that it compiles
*at all* is a testament to Rust and a joy to behold.

TODO: Write tests that examine each optimization in turn, especially those
that require reductions.

TODO: Consider splitting the reductions into an internal and external
pass, one that will allow the user to specify the data types that the
parse-null pass will extract, and then redefine the internal ones (the
tree-balancing ones) as data types with a common operator name, rather
than as closures.  Leave closures to the hackers.