REFACTOR Changed 'add_' to 'make_'
The internal representation here is not relevant to the operation;
we're making new nodes (where they get stored, 'add'ed to memory,
is an implementation detail), and we want the code to reflect that
change.

Changed 'panic!' to be 'unreachable!', as that's the actual
operation we're performing on a regular basis.