FEAT Float-after-reduction compliant with Might 2015
The 'Float-after-reduction' code was shimmed to do nothing
but work with the existing code; if it was composed with
user-defined reductions, everything broke.  This is now
complaint with Might 2015, but doesn't handle external
reductions.

In Might (2011), the section of reductions reads:

    A reduction by function f : X → Y over a parser p ∈ P(A, X)
    creates a new partial parser, p → f ∈ P(A, Y):

        p → f = λw.{((f(x), w') : (x, w')  ∈ p(w)}

    A reduction parser maps trees from X into trees from Y

The problem is that we now need a way to represent the trees of
type Y in our return set.  Until we figure that out, this project
is going to be nothing more than an academic exercise.