FEAT Epsilon data is now carried on epsilon nodes.
This eliminates one major small bit of confusion.  I've also
changed the null tracking inside the deriver into a hashmap for
the time being. This makes nullability a bit slower (every lookup
is O(log n), rather than O(1), but it's a heck of a lot easier to
keep track of.